refactor: tests move with their files
