+++ devtools/shell/src/utils/toolsUtils.test.ts
